utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
]
